Split the code that add listeners and process doms into modules
Update Makefile to use coqdocjs-static to preprocess the html files

Update Makefile and fix a typo in the function name

Add environment variables for coqdocjs static

Refactor to use for loops

Fix preprocess in index.js

Parallelize the preprocessing

Remove unneeded config.js file for static build

Add log messages

Typo

Fix bug in makefile

Revert back to the old efficient implementation

Update package lock

Remove the extra /
