+++ coqdocjs-static/README.md
@@ -1,0 +1,3 @@
+# coqdocjs-static
+A helper script that statically transforms the output of coqdoc so the
+expensive dom operations don't have to happen during run-time.
