+++ coqdocjs-static/preprocess.js
@@ -1,0 +1,1 @@
+../extra/resources/preprocess.js
