+++ extra/header.html
@@ -5,8 +5,7 @@
 <meta http-equiv="Content-Type" content="text/html;charset=utf-8" />
 <link href="resources/coqdoc.css" rel="stylesheet" type="text/css" />
 <link href="resources/coqdocjs.css" rel="stylesheet" type="text/css"/>
-<script type="text/javascript" src="resources/config.js"></script>
-<script type="text/javascript" src="resources/coqdocjs.js"></script>
+<script type="module" src="resources/coqdocjs.js"></script>
 </head>
 
 <body onload="document.getElementById('content').focus()">
